requirements.txt
# Web-Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Datenbank
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Datenvalidierung und Konfiguration
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# Authentifizierung und Sicherheit
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# HTTP und Netzwerk
requests==2.31.0

# Selenium und WebDriver
selenium==4.15.2
webdriver-manager==4.0.1
chromedriver-autoinstaller==0.4.0

# Logging und Monitoring
python-json-logger

# Hilfsprogramme
python-dateutil==2.8.2

# Entwicklungs- und Testwerkzeuge
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
coverage==7.3.2

# Performance
ujson==5.8.0
orjson==3.9.10

# Caching (optional)
redis==5.0.1